requirements.txt
torch~=2.5
torch_geometric~=2.5
networkx~=3.4
scikit-learn~=1.6
isort==6.0.1
ruff==0.11.8
mypy==1.15.0
click==8.1.8
dotenv==0.9.9
psutil==5.9.8
rich==13.9
hydra-core~=1.3
mlflow~=2.22
